updates save methods
